# api-spec.yaml
openapi: 3.0.0
servers: []
info:
  version: 0.4.6
  title: Reference Data MS
  description: ''
security:
  - accessCode:
      - read
      - write
...
paths:
...
  /ref-group/{ref_group_code}/ref_entries/:
    get:
      tags:
        - Customer
      summary: |
        Retrieve a set of Ref Entries under a Ref Group
      description: ''
      security: []
      parameters:
        - $ref: '#/components/parameters/Filter'
        - $ref: '#/components/parameters/RefGroupCode'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  Filter:
                    description: ''
                    type: string
                  Page:
                    description: ''
                    type: integer
                  PageSize:
                    description: ''
                    type: integer
                  Total:
                    description: ''
                    type: integer
                  Result:
                    type: array
                    items:
                      $ref: '#/components/schemas/RefEntryUser'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
  /admin/ref-groups:
    get:
      tags:
        - Admin
      summary: Retrieve all the existing groups.
      description: ''
      security: []
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RefGroup'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /admin/ref-group/{ref_group_code}:
    get:
      tags:
        - Admin
      summary: Retrieve only one RefGroup by its code with configuration details
      description: ''
      security: []
      parameters:
        - $ref: '#/components/parameters/RefGroupCode'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefGroup'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /admin/ref-group/{ref_group_code}/set:
    post:
      tags:
        - Admin
      summary: Change the configuration detail of a RefGroup, or create a RefGroup if the `code` doesn't exists.
      description: ''
      security: []
      parameters:
        - $ref: '#/components/parameters/RefGroupCode'
      requestBody:
        description: ''
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefGroup'
      responses:
        '200':
          $ref: '#/components/responses/200'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /admin/ref-group/{ref_group_code}/unset:
    post:
      tags:
        - Admin
      summary: Remove a RefGroup permenantly.
      description: ''
      security: []
      parameters:
        - $ref: '#/components/parameters/RefGroupCode'
      responses:
        '200':
          $ref: '#/components/responses/200'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /admin/ref-group/{ref_group_code}/ref-entries/:
    get:
      tags:
        - Admin
      summary: Retrieve all the RefEntries of a RefGroup
      description: ''
      security: []
      parameters:
        - $ref: '#/components/parameters/Filter'
        - $ref: '#/components/parameters/RefGroupCode'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  Page:
                    description: ''
                    type: integer
                  PageSize:
                    description: ''
                    type: integer
                  Total:
                    description: ''
                    type: integer
                  Result:
                    type: array
                    items:
                      $ref: '#/components/schemas/RefEntry'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /admin/ref-group/{ref_group_code}/ref-entry/{ref_entry_code}:
    get:
      tags:
        - Admin
      summary: Retrieve only one RefEntry with all its detail
      description: ''
      security: []
      parameters:
        - $ref: '#/components/parameters/RefGroupCode'
        - $ref: '#/components/parameters/RefEntryCode'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefEntryEditView'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /admin/ref-group/{ref_group_code}/ref-entry/{ref_entry_code}/set:
    post:
      tags:
        - Admin
      summary: Create
      description: ''
      security: []
      parameters:
        - $ref: '#/components/parameters/RefGroupCode'
        - $ref: '#/components/parameters/RefEntryCode'
      requestBody:
        description: ''
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefEntryEditViewSave'
      responses:
        '200':
          $ref: '#/components/responses/200'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /admin/ref-group/{ref_group_code}/ref-entry/{ref_entry_code}/unset:
    post:
      tags:
        - Admin
      summary: Remove an RefEntry identified by its `code`
      description: ''
      security: []
      parameters:
        - $ref: '#/components/parameters/RefGroupCode'
        - $ref: '#/components/parameters/RefEntryCode'
      responses:
        '200':
          $ref: '#/components/responses/200'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /admin/translations/:
    get:
      tags:
        - Admin
      summary: Retrieve all existing translation keys
      description: ''
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  Page:
                    description: ''
                    type: integer
                  PageSize:
                    description: ''
                    type: integer
                  Total:
                    description: ''
                    type: integer
                  Result:
                    type: array
                    items:
                      $ref: '#/components/schemas/Translation'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /admin/translation/{translation_key}/:
    get:
      tags:
        - Admin
      summary: Server heartbeat operation
      description: ''
      security: []
      parameters:
        - $ref: '#/components/parameters/TranslationKey'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TranslationEntry'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /admin/translation/{translation_key}/{locale}/set:
    post:
      tags:
        - Admin
      summary: Server heartbeat operation
      description: ''
      security: []
      parameters:
        - $ref: '#/components/parameters/TranslationKey'
        - $ref: '#/components/parameters/Locale'
      requestBody:
        description: ''
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationEntry'
      responses:
        '200':
          $ref: '#/components/responses/200'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /admin/translation/{translation_key}/{locale}/unset:
    post:
      tags:
        - Admin
      summary: Server heartbeat operation
      description: ''
      security: []
      parameters:
        - $ref: '#/components/parameters/TranslationKey'
        - $ref: '#/components/parameters/Locale'
      responses:
        '200':
          $ref: '#/components/responses/200'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
components:
  securitySchemes:
    accessCode:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: 'http://example.com/oauth/auth'
          tokenUrl: 'http://example.com/oauth/token'
          scopes:
            write: allows modifying resources
            read: allows reading resources
  parameters:
    Page:
      name: page
      in: query
      description: ''
      required: false
      schema:
        type: integer
    PageSize:
      name: page_size
      in: query
      description: ''
      required: false
      schema:
        type: integer
    RefGroupCode:
      in: path
      name: ref_group_code
      schema:
        type: string
      required: true
      description: ''
    RefEntryCode:
      in: path
      name: ref_entry_code
      schema:
        type: string
      required: true
      description: ''
    Locale:
      name: locale
      in: path
      description: ''
      required: true
      schema:
        type: string
    TranslationKey:
      name: translation_key
      in: path
      description: ''
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 200
    Filter:
      name: filter
      in: query
      description: ''
      required: false
      schema:
        type: string
  responses:
    200:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExtraInfo'
    400:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExtraInfo'
    500:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExtraInfo'
  schemas:
    ExtraInfo:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    Permissions:
      description: ''
      type: array
      items:
        type: string
        enum:
          - EDIT
          - REMOVE
    RefGroup:
      description: ''
      type: object
      properties:
        code:
          description: ''
          type: string
          minLength: 2
          maxLength: 32
        name:
          description: ''
          type: string
          minLength: 5
          maxLength: 200
        desc:
          description: ''
          type: string
          minLength: 0
          maxLength: 5000
        custom_schema:
          $ref: '#/components/schemas/JSONScheme'
        custom_forms:
          $ref: '#/components/schemas/JSONForms'
        allow:
          $ref: '#/components/schemas/Permissions'
      required:
        - name
        - desc
        - custom_schema
        - custom_forms
    RefEntry:
      description: ''
      type: object
      properties:
        ref_group_code:
          description: ''
          type: string
          minLength: 2
          maxLength: 32
        code:
          description: ''
          type: string
          minLength: 2
          maxLength: 32
        desc:
          description: ''
          type: string
          minLength: 0
          maxLength: 1000
        sec_id:
          description: ''
          type: integer
        custom_fields:
          $ref: '#/components/schemas/CustomFields'
        allow:
          $ref: '#/components/schemas/Permissions'
    RefEntryEditView:
      description: ''
      type: object
      properties:
        ui_schema:
          type: object
        custom_fields:
          $ref: '#/components/schemas/CustomFields'
      required:
        - ui_schema
        - custom_fields
    RefEntryEditViewSave:
      description: ''
      type: object
      properties:
        custom_fields:
          $ref: '#/components/schemas/CustomFields'
      required:
        - custom_fields
    RefEntryUser:
      description: ''
      type: object
      properties:
        code:
          description: ''
          type: string
          minLength: 2
          maxLength: 32
        data:
          $ref: '#/components/schemas/CustomFields'
      required:
        - code
        - data
    TranslationEntry:
      description: ''
      type: object
      properties:
        locale:
          description: ''
          type: string
        value:
          description: ''
          type: string
          minLength: 1
          maxLength: 10000
      required:
        - key
        - locale
        - value
    Translation:
      type: object
      properties:
        key:
          type: string
          minLength: 1
          maxLength: 200
        entries:
          type: array
          items:
            $ref: '#/components/schemas/TranslationEntry'
    JSONScheme:
      description: ''
      type: object
    JSONForms:
      description: ''
      type: object
    CustomFields:
      description: ''
      type: object
...